Create Temperature based Fan Speed Controller
A simple introduction about the temperature controller circuit. Day by day, there are different types of intelligent systems are introduced with the improvement in technology making our lives comfortable and less compact . Everything is getting more intelligible and technology oriented. There is a growth in the demand of cutting-edge technology and also smart electronic systems.
 In the proposed systems, microcontroller plays a vital role in the smart systems development. Microcontrollers have become an essential part in the present technologies that are being presented daily.
We aimed at making a temperature-based fan, in which there is a involvement of a  speed control and a monitoring system all embedded with Arduino board. This system is used to control the cooling function which is automatically based upon the room/surrounding temperature. The main component used is an Arduino board to implement a control system. Since this system is proposed to control the cooling structure thus it is very important to know Arduino controlled system quite well. 
This circuit project is mainly used for indicating temperature, controlling temperature. Microprocessor forms the processing part. The hardware circuit of temperature controller need the programming. The main principle of this project is to indicate temperature and to control the temperature.
The addon functionality of our project is the involvement of piezoelectric material which thereby acts as a source of energy conservation, and the fan will only be driven when certain mechanical forces bring about certain deformities in the state of piezoelectric material resulting in the generation of electric charge due to which our fan operates, automatically setting its speed in accordance with the surrounding temperature . As soon as the material gets back to its original state the charge will not be supplied to the circuit and the fan stops no matter what the temperature is.